refactor existing tests for UI REST Controllers, Service Activity and UI REST Controllers Trip add coverage for new cases

+++ src/main/java/io/github/valeriikuzmych/travelplanner/service/ActivityServiceImpl.java
@@ -112,6 +112,10 @@ public class ActivityServiceImpl implements ActivityService {
 
         ownershipValidator.assertUserOwnActivity(id, email);
 
+        if (!activityRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Activity not found");
+        }
+
         activityRepository.deleteById(id);
     }
